refactor(obd2): extract TripRecordingPhase + HapticFeedbackPolicy + TripRecordingState (Refs #563)
Split the 706-line trip_recording_provider into cohesive modules so
the provider itself stays focused on the Obd2Service lifecycle and
baseline bookkeeping. Three pulls, all re-exported from the
provider file so existing imports (widgets, screens, tests) keep
resolving unchanged:

- trip_recording_phase.dart — the TripRecordingPhase enum
- haptic_feedback_policy.dart — HapticIntensity + the pure
  hapticForBandTransition() function (now unit-tested exhaustively
  against every band-transition case to guard #767 behaviour)
- trip_recording_state.dart — the @immutable TripRecordingState
  DTO + copyWith

Provider shrinks from 706 → 639 LOC. No behaviour change.
